[오명섭] 자체빌드한 513*513 을 올려서 돌아가도록 만들었으나, 아무런 segmentation을 못하고 있다!

+++ android/app/src/main/java/org/tensorflow/demo/tracking/OverlayTracker.java
@@ -39,7 +39,7 @@ public class OverlayTracker {
   private Size previewSize;
 
   private int colors[];
-  private long[] resultPixels;
+  private byte[] resultPixels;
   private int[] pixels;
   private Vector<String> lastLabels;
   private Vector<String> labels;
